Update 0.1.6 Improved: * Mehod convertToASCII() code readability * Color structure new use short int instead of int

+++ src/UnixConsole.h
@@ -17,15 +17,15 @@ class UnixConsole
 private:
 	struct Color
 	{
-		int red;
-		int blue;
-		int green;
+		short int red;
+		short int blue;
+		short int green;
 	};
 	Color mainColor;
 public:
 	UnixConsole();
 	void setMainTextColor(Color a);
-	Color setColor(int  r, int g, int b);
+	Color setColor(short int r, short int g, short int b);
 	void setColorText(Color a, const char* text);
 	void writeText(std::string text);
 	void setColorText(int  r, int g, int b, const char* text);
